requirements.txt
absl-py==2.2.1
asttokens==3.0.0
astunparse==1.6.3
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725571112467/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
comm==0.2.2
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2024.6.1
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.13.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata==8.6.1
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459266648/work
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1659031536286/work
matplotlib-inline==0.1.7
mpmath==1.3.0
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806303256/work
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==1.3.5
parso==0.8.4
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pexpect==4.9.0
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929703139/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1742485085/work
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1733239724146/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1733421311631/work
prompt_toolkit==3.0.50
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
Pygments==2.19.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PyQt5==5.15.10
PyQt5_sip==12.13.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
pyzmq==26.3.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-oauthlib==2.0.0
rsa==4.9
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1692501296392/work/base/dist/scipy-1.11.2-cp39-cp39-linux_x86_64.whl#sha256=3b1a8484ce148b8f528582054e76a2a71f677f9b8ed2c1871547ec805e38aa03
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1629095986539/work
sip @ file:///home/conda/feedstock_root/build_artifacts/sip_1742163291223/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
stack-data==0.6.3
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1683305560313/work
sympy==1.13.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1733256695513/work
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615921868/work
traitlets==5.14.3
triton==3.3.0
typing_extensions==4.12.2
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692503055/work
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
